services/instance/requirements.txt
# FastAPI and server
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis
redis==5.0.1
hiredis==2.3.2

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Validation
pydantic==2.7.4
pydantic-settings==2.3.2
email-validator==2.1.0

# Authentication (JWT)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Utilities
python-dotenv==1.0.1
python-json-logger==2.0.7

# Date/Time
python-dateutil==2.8.2
